requirements.txt
Flask
Flask-Cors
torch
timm
albumentations
Pillow
numpy
opencv-python-headless
gunicorn